% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casos_aplicados.R
\name{print.rnas_caso_aplicado_fnl}
\alias{print.rnas_caso_aplicado_fnl}
\title{Imprimir caso aplicado FNL-RNAS}
\usage{
\method{print}{rnas_caso_aplicado_fnl}(x, ...)
}
\arguments{
\item{x}{Objeto `rnas_caso_aplicado_fnl`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto x de forma invisible.
}
\description{
Imprimir caso aplicado FNL-RNAS
}
