% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casos_aplicados.R
\name{print.rnas_caso_aplicado_simuds}
\alias{print.rnas_caso_aplicado_simuds}
\title{Imprimir caso aplicado SimuDS-RNAS}
\usage{
\method{print}{rnas_caso_aplicado_simuds}(x, ...)
}
\arguments{
\item{x}{Objeto `rnas_caso_aplicado_simuds`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto x de forma invisible.
}
\description{
Imprimir caso aplicado SimuDS-RNAS
}
