% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casos_aplicados.R
\name{print.rnas_caso_aplicado_std2}
\alias{print.rnas_caso_aplicado_std2}
\title{Imprimir caso aplicado STD2-RNAS}
\usage{
\method{print}{rnas_caso_aplicado_std2}(x, ...)
}
\arguments{
\item{x}{Objeto `rnas_caso_aplicado_std2`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto x de forma invisible.
}
\description{
Imprimir caso aplicado STD2-RNAS
}
