% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casos_aplicados.R
\name{print.rnas_casos_aplicados}
\alias{print.rnas_casos_aplicados}
\title{Imprimir ejecución integral de casos aplicados}
\usage{
\method{print}{rnas_casos_aplicados}(x, ...)
}
\arguments{
\item{x}{Objeto `rnas_casos_aplicados`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto x de forma invisible.
}
\description{
Imprimir ejecución integral de casos aplicados
}
