% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_eta.R
\name{print.rnas_control_eta_train}
\alias{print.rnas_control_eta_train}
\title{Imprimir entrenamiento con control de tasa RNAS}
\usage{
\method{print}{rnas_control_eta_train}(x, ...)
}
\arguments{
\item{x}{Objeto de clase `rnas_control_eta_train`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto `x`, de forma invisible.
}
\description{
Metodo de impresion para objetos `rnas_control_eta_train`.
}
