% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integracion.R
\name{print.rnas_fnl_formulation}
\alias{print.rnas_fnl_formulation}
\title{Imprimir formulación FNL-RNAS}
\usage{
\method{print}{rnas_fnl_formulation}(x, ...)
}
\arguments{
\item{x}{Objeto `rnas_fnl_formulation`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto x de forma invisible.
}
\description{
Imprimir formulación FNL-RNAS
}
