% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_optimo.R
\name{print.rnas_meta_geo_train}
\alias{print.rnas_meta_geo_train}
\title{Imprimir entrenamiento con meta-control geometrico RNAS}
\usage{
\method{print}{rnas_meta_geo_train}(x, ...)
}
\arguments{
\item{x}{Objeto de clase `rnas_meta_geo_train`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto `x`, de forma invisible.
}
\description{
Imprimir entrenamiento con meta-control geometrico RNAS
}
