% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_optimo.R
\name{print.rnas_meta_policy_comparison}
\alias{print.rnas_meta_policy_comparison}
\title{Imprimir comparacion de politicas de meta-control RNAS}
\usage{
\method{print}{rnas_meta_policy_comparison}(x, ...)
}
\arguments{
\item{x}{Objeto de clase `rnas_meta_policy_comparison`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto `x`, de forma invisible.
}
\description{
Imprimir comparacion de politicas de meta-control RNAS
}
