% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{print.rnas_mlp_train}
\alias{print.rnas_mlp_train}
\title{Imprimir resumen de entrenamiento de MLP RNAS}
\usage{
\method{print}{rnas_mlp_train}(x, ...)
}
\arguments{
\item{x}{Objeto de clase `rnas_mlp_train`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto `x`, de forma invisible.
}
\description{
Metodo de impresion para objetos creados con `rnas_train_mlp()`.
}
