% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrenamiento_neurona.R
\name{print.rnas_neuron_train}
\alias{print.rnas_neuron_train}
\title{Imprimir resumen de entrenamiento de neurona RNAS}
\usage{
\method{print}{rnas_neuron_train}(x, ...)
}
\arguments{
\item{x}{Objeto de clase `rnas_neuron_train`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto `x`, de forma invisible.
}
\description{
Método de impresión para objetos creados con `rnas_train_neuron()`.
}
