% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integracion.R
\name{print.rnas_simuds_integration}
\alias{print.rnas_simuds_integration}
\title{Imprimir integración SimuDS-RNAS}
\usage{
\method{print}{rnas_simuds_integration}(x, ...)
}
\arguments{
\item{x}{Objeto `rnas_simuds_integration`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto x de forma invisible.
}
\description{
Imprimir integración SimuDS-RNAS
}
