% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integracion.R
\name{print.rnas_std2_features}
\alias{print.rnas_std2_features}
\title{Imprimir objeto de preparación STD2 para RNAS}
\usage{
\method{print}{rnas_std2_features}(x, ...)
}
\arguments{
\item{x}{Objeto `rnas_std2_features`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto x de forma invisible.
}
\description{
Imprimir objeto de preparación STD2 para RNAS
}
