% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integracion.R
\name{print.rnas_std2_integration}
\alias{print.rnas_std2_integration}
\title{Imprimir integración STD2-RNAS}
\usage{
\method{print}{rnas_std2_integration}(x, ...)
}
\arguments{
\item{x}{Objeto `rnas_std2_integration`.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
El objeto x de forma invisible.
}
\description{
Imprimir integración STD2-RNAS
}
