% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_eta.R
\name{rnas_clip_eta}
\alias{rnas_clip_eta}
\title{Aplicar cotas a una tasa de aprendizaje RNAS}
\usage{
rnas_clip_eta(eta, eta_min = 1e-05, eta_max = 1)
}
\arguments{
\item{eta}{Escalar numerico de tasa de aprendizaje.}

\item{eta_min}{Cota inferior positiva.}

\item{eta_max}{Cota superior positiva.}
}
\value{
Tasa de aprendizaje acotada.
}
\description{
Controla que una tasa de aprendizaje permanezca dentro de los limites
definidos por eta_min y eta_max.
}
\examples{
rnas_clip_eta(eta = 0.5, eta_min = 0.01, eta_max = 0.1)

}
