% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_optimo.R
\name{rnas_comparar_meta_politicas}
\alias{rnas_comparar_meta_politicas}
\title{Comparar politicas de meta-control geometrico RNAS}
\usage{
rnas_comparar_meta_politicas(
  X,
  y,
  theta0,
  politicas,
  eta0 = 0.1,
  eta_min = 0.01,
  eta_max = 0.1,
  T = 100,
  activation = "tanh"
)
}
\arguments{
\item{X}{Matriz numerica de entradas.}

\item{y}{Vector numerico de valores observados.}

\item{theta0}{Vector inicial de estado.}

\item{politicas}{Lista nombrada de configuraciones. Cada elemento debe
contener al menos `metodo`.}

\item{eta0}{Tasa base.}

\item{eta_min}{Cota inferior.}

\item{eta_max}{Cota superior.}

\item{T}{Numero de iteraciones.}

\item{activation}{Activacion.}
}
\value{
Objeto de clase `rnas_meta_policy_comparison`.
}
\description{
Ejecuta varias politicas de meta-control geometrico sobre el mismo caso
y devuelve una tabla comparativa.
}
