% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_eta.R
\name{rnas_comparar_politicas_eta}
\alias{rnas_comparar_politicas_eta}
\title{Comparar politicas de tasa de aprendizaje RNAS}
\usage{
rnas_comparar_politicas_eta(
  X,
  y,
  theta0,
  politicas,
  eta0 = 0.1,
  eta_min = 1e-05,
  eta_max = 1,
  T = 100,
  activation = "tanh"
)
}
\arguments{
\item{X}{Matriz numerica de entradas.}

\item{y}{Vector numerico de valores observados.}

\item{theta0}{Vector inicial con pesos y sesgo.}

\item{politicas}{Lista nombrada de politicas. Cada elemento puede ser texto
o lista con campo `tipo`.}

\item{eta0}{Tasa inicial.}

\item{eta_min}{Cota inferior de tasa.}

\item{eta_max}{Cota superior de tasa.}

\item{T}{Numero de iteraciones.}

\item{activation}{Activacion de la neurona.}
}
\value{
Objeto de clase `rnas_eta_policy_comparison`.
}
\description{
Ejecuta varias politicas de control de tasa de aprendizaje sobre el mismo
caso y devuelve una tabla comparativa de metricas.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
theta0 <- c(0.8, 0.3, 0.1)
politicas <- list(base = "constante", temporal = list(tipo = "temporal"))
comp <- rnas_comparar_politicas_eta(X, y, theta0, politicas, T = 20)
comp$comparacion

}
