% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casos_aplicados.R
\name{rnas_consolidar_casos}
\alias{rnas_consolidar_casos}
\title{Consolidar casos aplicados RNAS}
\usage{
rnas_consolidar_casos(casos)
}
\arguments{
\item{casos}{Lista de objetos de casos aplicados.}
}
\value{
Data frame consolidado.
}
\description{
Consolida las métricas principales de una lista de casos aplicados.
}
\examples{
c1 <- rnas_caso_std2_controlado(n = 30, T = 10)
c2 <- rnas_caso_simuds_controlado(n = 20, T = 10)
rnas_consolidar_casos(list(c1, c2))

}
