% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometria_aprendizaje.R
\name{rnas_curvatura_direccional}
\alias{rnas_curvatura_direccional}
\title{Calcular curvatura direccional}
\usage{
rnas_curvatura_direccional(H, v)
}
\arguments{
\item{H}{Matriz Hessiana numerica cuadrada.}

\item{v}{Vector numerico de direccion.}
}
\value{
Escalar numerico con la curvatura direccional.
}
\description{
Calcula la curvatura direccional v^T H v, normalizando la direccion
si no tiene norma unitaria.
}
\examples{
H <- diag(c(1, 2))
rnas_curvatura_direccional(H, v = c(1, 1))

}
