% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activaciones.R
\name{rnas_dsigmoid}
\alias{rnas_dsigmoid}
\title{Derivada de la activacion sigmoide RNAS}
\usage{
rnas_dsigmoid(z)
}
\arguments{
\item{z}{Valor numerico, vector numerico o matriz numerica.}
}
\value{
Objeto numerico del mismo tipo dimensional que `z`, con los valores de la derivada.
}
\description{
Evalua la derivada de la funcion sigmoide. La derivada se calcula como:
sigma'(z) = sigma(z) * (1 - sigma(z)).
}
\examples{
rnas_dsigmoid(0)
rnas_dsigmoid(c(-2, 0, 2))

}
