% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activaciones.R
\name{rnas_dtanh}
\alias{rnas_dtanh}
\title{Derivada de la activacion tangente hiperbolica RNAS}
\usage{
rnas_dtanh(z)
}
\arguments{
\item{z}{Valor numerico, vector numerico o matriz numerica.}
}
\value{
Objeto numerico del mismo tipo dimensional que `z`, con los valores de la derivada.
}
\description{
Evalua la derivada de la tangente hiperbolica. La derivada se calcula como:
tanh'(z) = 1 - tanh(z)^2.
}
\examples{
rnas_dtanh(0)
rnas_dtanh(c(-2, 0, 2))

}
