% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_optimo.R
\name{rnas_eta_costo_local}
\alias{rnas_eta_costo_local}
\title{Calcular costo local para una tasa candidata RNAS}
\usage{
rnas_eta_costo_local(
  X,
  y,
  theta,
  grad,
  eta,
  activation = "tanh",
  alpha = 0,
  beta = 0
)
}
\arguments{
\item{X}{Matriz numerica de entradas.}

\item{y}{Vector numerico de valores observados.}

\item{theta}{Vector numerico de estado actual.}

\item{grad}{Vector numerico de gradiente actual.}

\item{eta}{Tasa candidata positiva.}

\item{activation}{Nombre de activacion. Puede ser `"tanh"` o `"sigmoid"`.}

\item{alpha}{Penalizacion por magnitud de eta.}

\item{beta}{Penalizacion por desplazamiento parametrico.}
}
\value{
Lista con costo total, perdida tentativa, penalizaciones y theta_plus.
}
\description{
Evalua el costo local asociado a una tasa de aprendizaje candidata.
Para una tasa eta, se calcula una actualizacion tentativa:
theta_plus = theta - eta * grad,
luego se evalua la perdida tentativa y penalizaciones por tasa y desplazamiento.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
theta <- c(0.8, 0.3, 0.1)
grad <- c(-0.0007526453, -0.0101555834, -0.0114716749)
rnas_eta_costo_local(X, y, theta, grad, eta = 0.1)

}
