% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_optimo.R
\name{rnas_eta_opt_local}
\alias{rnas_eta_opt_local}
\title{Seleccionar tasa optima local por busqueda discreta RNAS}
\usage{
rnas_eta_opt_local(
  X,
  y,
  theta,
  grad,
  eta_grid,
  activation = "tanh",
  alpha = 0,
  beta = 0
)
}
\arguments{
\item{X}{Matriz numerica de entradas.}

\item{y}{Vector numerico de valores observados.}

\item{theta}{Vector numerico de estado actual.}

\item{grad}{Vector numerico de gradiente actual.}

\item{eta_grid}{Vector numerico de tasas candidatas.}

\item{activation}{Nombre de activacion.}

\item{alpha}{Penalizacion por magnitud de eta.}

\item{beta}{Penalizacion por desplazamiento parametrico.}
}
\value{
Lista con eta seleccionada, costo minimo y tabla de costos.
}
\description{
Evalua una grilla de tasas candidatas y selecciona la que minimiza
el costo local.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
theta <- c(0.8, 0.3, 0.1)
grad <- c(-0.0007526453, -0.0101555834, -0.0114716749)
rnas_eta_opt_local(X, y, theta, grad, eta_grid = c(0.01, 0.05, 0.1))

}
