% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_continua.R
\name{rnas_eval_eta}
\alias{rnas_eval_eta}
\title{Evaluar tasa de aprendizaje dinamica RNAS}
\usage{
rnas_eval_eta(eta, t, theta = NULL, loss = NULL, grad_norm = NULL)
}
\arguments{
\item{eta}{Escalar numerico positivo o funcion.}

\item{t}{Tiempo actual.}

\item{theta}{Estado actual.}

\item{loss}{Perdida actual.}

\item{grad_norm}{Norma del gradiente actual.}
}
\value{
Escalar numerico positivo.
}
\description{
Evalua una tasa de aprendizaje que puede ser escalar numerico o funcion.
Si es funcion, debe aceptar al menos el argumento `t`.
}
\examples{
rnas_eval_eta(0.1, t = 0)
rnas_eval_eta(function(t, ...) 0.1 / (1 + t), t = 2)

}
