% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activaciones.R
\name{rnas_get_activation}
\alias{rnas_get_activation}
\title{Obtener activacion RNAS por nombre}
\usage{
rnas_get_activation(nombre = c("sigmoid", "tanh"))
}
\arguments{
\item{nombre}{Cadena de texto con el nombre de la activacion.
Puede ser `"sigmoid"` o `"tanh"`.}
}
\value{
Lista con dos elementos:
\describe{
  \item{activation}{Funcion de activacion.}
  \item{derivative}{Derivada de la activacion.}
}
}
\description{
Devuelve una lista con la funcion de activacion y su derivada asociada.
Esta funcion estandariza el uso de activaciones dentro del paquete.
}
\examples{
act <- rnas_get_activation("tanh")
act$activation(0)
act$derivative(0)

}
