% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perdida_gradiente.R
\name{rnas_grad_neuron}
\alias{rnas_grad_neuron}
\title{Calcular gradiente analítico para una neurona RNAS}
\usage{
rnas_grad_neuron(X, y, w, b, activation = "tanh")
}
\arguments{
\item{X}{Matriz numérica de entradas.}

\item{y}{Vector numérico de valores observados.}

\item{w}{Vector numérico de pesos.}

\item{b}{Sesgo numérico escalar.}

\item{activation}{Nombre de activación. Puede ser `"sigmoid"` o `"tanh"`.}
}
\value{
Lista con:
\describe{
  \item{grad_w}{Vector de gradientes respecto a los pesos.}
  \item{grad_b}{Gradiente respecto al sesgo.}
  \item{loss}{Valor de pérdida MSE.}
  \item{y_hat}{Predicciones de la neurona.}
  \item{error}{Errores de predicción.}
  \item{z}{Potenciales internos.}
}
}
\description{
Calcula el gradiente analítico de la pérdida MSE respecto a los pesos
y al sesgo de una neurona individual.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8)
w <- c(0.8, 0.3)
b <- 0.1
rnas_grad_neuron(X, y, w, b, activation = "tanh")

}
