% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometria_aprendizaje.R
\name{rnas_hessian_num_neuron}
\alias{rnas_hessian_num_neuron}
\title{Calcular Hessiano numerico para una neurona RNAS}
\usage{
rnas_hessian_num_neuron(X, y, theta, activation = "tanh", h = 1e-04)
}
\arguments{
\item{X}{Matriz numerica de entradas.}

\item{y}{Vector numerico de valores observados.}

\item{theta}{Vector numerico de estado con pesos y sesgo.}

\item{activation}{Nombre de activacion. Puede ser `"tanh"` o `"sigmoid"`.}

\item{h}{Paso positivo para diferencias finitas.}
}
\value{
Matriz Hessiana aproximada.
}
\description{
Aproxima numericamente la matriz Hessiana de la perdida MSE respecto
al vector de estado theta = c(w, b), usando diferencias finitas centrales.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
theta <- c(0.8, 0.3, 0.1)
H <- rnas_hessian_num_neuron(X, y, theta)
H

}
