% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{rnas_init_mlp}
\alias{rnas_init_mlp}
\title{Inicializar un MLP simple RNAS}
\usage{
rnas_init_mlp(d_input, d_hidden, init_sd = 0.1, seed = NULL)
}
\arguments{
\item{d_input}{Entero positivo. Numero de variables de entrada.}

\item{d_hidden}{Entero positivo. Numero de neuronas ocultas.}

\item{init_sd}{Desviacion estandar para inicializar pesos aleatorios.}

\item{seed}{Semilla opcional para reproducibilidad.}
}
\value{
Lista con parametros `W`, `b1`, `v` y `b2`.
}
\description{
Inicializa los parametros de un perceptron multicapa simple con una capa
oculta. La arquitectura considerada es:
y_hat = v^T phi(Wx + b1) + b2.
}
\examples{
params <- rnas_init_mlp(d_input = 2, d_hidden = 3, seed = 123)
str(params)

}
