% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometria_aprendizaje.R
\name{rnas_loss_grid_neuron}
\alias{rnas_loss_grid_neuron}
\title{Construir malla de perdida para una neurona RNAS}
\usage{
rnas_loss_grid_neuron(X, y, w1_seq, w2_seq, b, activation = "tanh")
}
\arguments{
\item{X}{Matriz numerica de entradas. Debe tener al menos dos columnas.}

\item{y}{Vector numerico de valores observados.}

\item{w1_seq}{Vector numerico con valores para el primer peso.}

\item{w2_seq}{Vector numerico con valores para el segundo peso.}

\item{b}{Sesgo numerico fijo.}

\item{activation}{Nombre de activacion. Puede ser `"tanh"` o `"sigmoid"`.}
}
\value{
Data frame con columnas `w1`, `w2` y `loss`.
}
\description{
Construye una malla de valores de perdida variando dos pesos de una neurona
y manteniendo fijo el sesgo. Esta funcion permite representar el paisaje
de perdida en dos dimensiones.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
grid <- rnas_loss_grid_neuron(X, y,
                              w1_seq = seq(0.7, 0.9, length.out = 5),
                              w2_seq = seq(0.2, 0.4, length.out = 5),
                              b = 0.1)
head(grid)

}
