% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perdida_gradiente.R
\name{rnas_loss_mse_neuron}
\alias{rnas_loss_mse_neuron}
\title{Calcular pérdida MSE para una neurona RNAS}
\usage{
rnas_loss_mse_neuron(X, y, w, b, activation = "tanh", devolver_detalle = FALSE)
}
\arguments{
\item{X}{Matriz numérica de entradas.}

\item{y}{Vector numérico de valores observados.}

\item{w}{Vector numérico de pesos.}

\item{b}{Sesgo numérico escalar.}

\item{activation}{Función de activación o nombre de activación.
Puede ser `"sigmoid"` o `"tanh"`.}

\item{devolver_detalle}{Valor lógico. Si es `TRUE`, retorna predicciones,
errores y pérdida.}
}
\value{
Si `devolver_detalle = FALSE`, retorna la pérdida MSE como escalar.
Si `devolver_detalle = TRUE`, retorna una lista con `loss`, `y_hat`,
`error` y `activation`.
}
\description{
Calcula la pérdida cuadrática media entre los valores observados `y`
y las predicciones generadas por una neurona individual.
}
\details{
La pérdida se define como:
L = mean((y_hat - y)^2).
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8)
w <- c(0.8, 0.3)
b <- 0.1
rnas_loss_mse_neuron(X, y, w, b, activation = "tanh")

}
