% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{rnas_mlp_backward}
\alias{rnas_mlp_backward}
\title{Retropropagacion para un MLP simple RNAS}
\usage{
rnas_mlp_backward(X, y, params, activation = "tanh")
}
\arguments{
\item{X}{Matriz numerica de entradas.}

\item{y}{Vector numerico de valores observados.}

\item{params}{Lista con parametros `W`, `b1`, `v` y `b2`.}

\item{activation}{Nombre de activacion. Puede ser `"tanh"` o `"sigmoid"`.}
}
\value{
Lista con gradientes, perdida, predicciones y errores.
}
\description{
Calcula los gradientes analiticos de la perdida MSE respecto a todos los
parametros del MLP simple.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8)
params <- rnas_init_mlp(d_input = 2, d_hidden = 3, seed = 123)
rnas_mlp_backward(X, y, params)

}
