% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{rnas_mlp_forward}
\alias{rnas_mlp_forward}
\title{Propagacion hacia adelante en un MLP simple RNAS}
\usage{
rnas_mlp_forward(X, params, activation = "tanh", devolver_cache = FALSE)
}
\arguments{
\item{X}{Matriz numerica de entradas.}

\item{params}{Lista con parametros `W`, `b1`, `v` y `b2`.}

\item{activation}{Nombre de activacion. Puede ser `"tanh"` o `"sigmoid"`.}

\item{devolver_cache}{Valor logico. Si es `TRUE`, devuelve tambien `Z1` y `H`.}
}
\value{
Si `devolver_cache = FALSE`, retorna vector de predicciones.
Si `TRUE`, retorna lista con `Z1`, `H`, `y_hat` y `activation`.
}
\description{
Calcula potenciales ocultos, activaciones ocultas y predicciones de un
perceptron multicapa simple con una capa oculta.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3), ncol = 2, byrow = TRUE)
params <- rnas_init_mlp(d_input = 2, d_hidden = 3, seed = 123)
rnas_mlp_forward(X, params, activation = "tanh")

}
