% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurona.R
\name{rnas_neuron_forward}
\alias{rnas_neuron_forward}
\title{Evaluar una neurona individual RNAS}
\usage{
rnas_neuron_forward(x, w, b, activation = "tanh", devolver_z = FALSE)
}
\arguments{
\item{x}{Vector numerico de entrada.}

\item{w}{Vector numerico de pesos.}

\item{b}{Sesgo numerico escalar.}

\item{activation}{Funcion de activacion o nombre de activacion.
Puede ser una funcion R o una cadena: `"sigmoid"` o `"tanh"`.}

\item{devolver_z}{Valor logico. Si es `TRUE`, devuelve tambien el potencial interno `z`.}
}
\value{
Si `devolver_z = FALSE`, retorna la salida estimada `y_hat`.
Si `devolver_z = TRUE`, retorna una lista con `z`, `y_hat` y `activation`.
}
\description{
Calcula el potencial interno y la salida estimada de una neurona individual.
La neurona se define como:
y_hat = phi(w^T x + b).
}
\examples{
x <- c(0.5, -0.2)
w <- c(0.8, 0.3)
b <- 0.1

rnas_neuron_forward(x, w, b, activation = "tanh")
rnas_neuron_forward(x, w, b, activation = "tanh", devolver_z = TRUE)

}
