% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_continua.R
\name{rnas_pack_neuron_params}
\alias{rnas_pack_neuron_params}
\title{Empaquetar parametros de una neurona RNAS}
\usage{
rnas_pack_neuron_params(w, b)
}
\arguments{
\item{w}{Vector numerico de pesos.}

\item{b}{Sesgo numerico escalar.}
}
\value{
Vector numerico theta = c(w, b).
}
\description{
Agrupa pesos y sesgo en un unico vector de estado theta.
Si w tiene longitud d, entonces theta tiene longitud d + 1.
}
\examples{
theta <- rnas_pack_neuron_params(w = c(0.8, 0.3), b = 0.1)
theta

}
