% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_continua.R
\name{rnas_predict_dinamica_neuron}
\alias{rnas_predict_dinamica_neuron}
\title{Predecir desde la dinamica de una neurona RNAS}
\usage{
rnas_predict_dinamica_neuron(object, X)
}
\arguments{
\item{object}{Objeto de clase `rnas_neuron_dynamics`.}

\item{X}{Matriz numerica de entradas.}
}
\value{
Vector numerico de predicciones.
}
\description{
Genera predicciones usando el estado final de una trayectoria dinamica.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8)
theta0 <- rnas_pack_neuron_params(c(0.8, 0.3), 0.1)
res <- rnas_integrar_dinamica_neuron(X, y, theta0, eta = 0.1, dt = 1, T = 10)
rnas_predict_dinamica_neuron(res, X)

}
