% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{rnas_predict_mlp}
\alias{rnas_predict_mlp}
\title{Predecir con un MLP simple RNAS entrenado}
\usage{
rnas_predict_mlp(object = NULL, X, params = NULL, activation = "tanh")
}
\arguments{
\item{object}{Objeto `rnas_mlp_train` o `NULL`.}

\item{X}{Matriz numerica de entradas.}

\item{params}{Lista de parametros. Se usa si `object = NULL`.}

\item{activation}{Activacion. Se usa si `object = NULL`.}
}
\value{
Vector numerico de predicciones.
}
\description{
Genera predicciones usando un objeto entrenado por `rnas_train_mlp()`
o usando parametros proporcionados directamente.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3), ncol = 2, byrow = TRUE)
params <- rnas_init_mlp(2, 3, seed = 123)
rnas_predict_mlp(NULL, X, params = params)

}
