% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrenamiento_neurona.R
\name{rnas_predict_neuron}
\alias{rnas_predict_neuron}
\title{Predecir con una neurona RNAS entrenada}
\usage{
rnas_predict_neuron(object = NULL, X, w = NULL, b = NULL, activation = "tanh")
}
\arguments{
\item{object}{Objeto `rnas_neuron_train` o `NULL`.}

\item{X}{Matriz numérica de entradas.}

\item{w}{Vector de pesos. Se usa si `object = NULL`.}

\item{b}{Sesgo numérico. Se usa si `object = NULL`.}

\item{activation}{Activación. Se usa si `object = NULL`.}
}
\value{
Vector numérico de predicciones.
}
\description{
Genera predicciones usando un objeto entrenado por `rnas_train_neuron()`
o usando pesos y sesgo proporcionados directamente.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3), ncol = 2, byrow = TRUE)
rnas_predict_neuron(NULL, X, w = c(0.8, 0.3), b = 0.1)

}
