% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_eta.R
\name{rnas_resumen_control_eta}
\alias{rnas_resumen_control_eta}
\title{Resumir entrenamiento con control de tasa RNAS}
\usage{
rnas_resumen_control_eta(object, nombre_politica = NULL)
}
\arguments{
\item{object}{Objeto de clase `rnas_control_eta_train`.}

\item{nombre_politica}{Nombre opcional de la politica.}
}
\value{
Data frame con metricas principales.
}
\description{
Genera una tabla resumen de un objeto producido por
`rnas_train_neuron_control_eta()`.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
theta0 <- c(0.8, 0.3, 0.1)
res <- rnas_train_neuron_control_eta(X, y, theta0, T = 20)
rnas_resumen_control_eta(res)

}
