% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_continua.R
\name{rnas_resumen_dinamica_neuron}
\alias{rnas_resumen_dinamica_neuron}
\title{Resumir dinamica continua de una neurona RNAS}
\usage{
rnas_resumen_dinamica_neuron(object)
}
\arguments{
\item{object}{Objeto de clase `rnas_neuron_dynamics`.}
}
\value{
Data frame con metricas principales de la dinamica.
}
\description{
Genera una tabla resumen de la trayectoria dinamica producida por
`rnas_integrar_dinamica_neuron()`.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
theta0 <- rnas_pack_neuron_params(c(0.8, 0.3), 0.1)
res <- rnas_integrar_dinamica_neuron(X, y, theta0, eta = 0.1, dt = 1, T = 50)
rnas_resumen_dinamica_neuron(res)

}
