% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{rnas_resumen_entrenamiento_mlp}
\alias{rnas_resumen_entrenamiento_mlp}
\title{Resumir entrenamiento de un MLP simple RNAS}
\usage{
rnas_resumen_entrenamiento_mlp(object)
}
\arguments{
\item{object}{Objeto de clase `rnas_mlp_train`.}
}
\value{
Data frame con resumen del entrenamiento.
}
\description{
Genera una tabla resumen con las metricas principales de un entrenamiento
realizado con `rnas_train_mlp()`.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
res <- rnas_train_mlp(X, y, d_hidden = 3, eta = 0.1, T = 50, seed = 123)
rnas_resumen_entrenamiento_mlp(res)

}
