% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrenamiento_neurona.R
\name{rnas_resumen_entrenamiento_neuron}
\alias{rnas_resumen_entrenamiento_neuron}
\title{Resumir entrenamiento de una neurona RNAS}
\usage{
rnas_resumen_entrenamiento_neuron(object)
}
\arguments{
\item{object}{Objeto de clase `rnas_neuron_train`.}
}
\value{
Data frame con pérdida inicial, pérdida final, reducción,
iteración final, criterio de parada y configuración principal.
}
\description{
Genera una tabla resumen con métricas principales de un entrenamiento
realizado con `rnas_train_neuron()`.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3,
             -0.7,  0.8,
              0.0,  0.0), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8, -0.2, 0.1)
res <- rnas_train_neuron(X, y, w0 = c(0.8, 0.3), b0 = 0.1,
                         eta = 0.1, T = 50)
rnas_resumen_entrenamiento_neuron(res)

}
