% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integracion.R
\name{rnas_resumen_integracion}
\alias{rnas_resumen_integracion}
\title{Resumir integración DS Neural RNAS}
\usage{
rnas_resumen_integracion(object)
}
\arguments{
\item{object}{Objeto de integración RNAS.}
}
\value{
Data frame con resumen interpretativo.
}
\description{
Consolida un resumen general de objetos de integración STD2, SimuDS o FNL.
}
