% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimenes.R
\name{rnas_segmentar_regimenes}
\alias{rnas_segmentar_regimenes}
\title{Segmentar regimenes dinamicos consecutivos}
\usage{
rnas_segmentar_regimenes(
  trayectoria_regimen,
  iter_col = "iter",
  loss_col = "loss",
  grad_col = "grad_norm",
  speed_col = "speed"
)
}
\arguments{
\item{trayectoria_regimen}{Data frame con columna `regimen`.}

\item{iter_col}{Nombre de la columna de iteracion.}

\item{loss_col}{Nombre de la columna de perdida.}

\item{grad_col}{Nombre de la columna de gradiente.}

\item{speed_col}{Nombre de la columna de velocidad.}
}
\value{
Data frame con segmentos consecutivos de regimen.
}
\description{
Agrupa tramos consecutivos de una trayectoria que comparten la misma
etiqueta de regimen.
}
\examples{
df <- data.frame(
  iter = 0:4,
  loss = c(1, .8, .7, .69, .68),
  grad_norm = c(.5, .4, .3, .2, .1),
  speed = c(.05, .04, .03, .02, .01),
  regimen = c("A", "B", "B", "C", "C")
)
rnas_segmentar_regimenes(df)

}
