% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activaciones.R
\name{rnas_sigmoid}
\alias{rnas_sigmoid}
\title{Activacion sigmoide RNAS}
\usage{
rnas_sigmoid(z)
}
\arguments{
\item{z}{Valor numerico, vector numerico o matriz numerica.}
}
\value{
Objeto numerico del mismo tipo dimensional que `z`, con valores entre 0 y 1.
}
\description{
Evalua la funcion sigmoide sobre un valor escalar, vector o matriz numerica.
La funcion sigmoide se define como:
sigma(z) = 1 / (1 + exp(-z)).
}
\examples{
rnas_sigmoid(0)
rnas_sigmoid(c(-2, 0, 2))

}
