% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activaciones.R
\name{rnas_tanh}
\alias{rnas_tanh}
\title{Activacion tangente hiperbolica RNAS}
\usage{
rnas_tanh(z)
}
\arguments{
\item{z}{Valor numerico, vector numerico o matriz numerica.}
}
\value{
Objeto numerico del mismo tipo dimensional que `z`, con valores entre -1 y 1.
}
\description{
Evalua la funcion tangente hiperbolica sobre un valor escalar,
vector o matriz numerica.
}
\examples{
rnas_tanh(0)
rnas_tanh(c(-2, 0, 2))

}
