% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_continua.R
\name{rnas_unpack_neuron_params}
\alias{rnas_unpack_neuron_params}
\title{Desempaquetar parametros de una neurona RNAS}
\usage{
rnas_unpack_neuron_params(theta, d_input)
}
\arguments{
\item{theta}{Vector numerico de estado.}

\item{d_input}{Entero positivo con la dimension de entrada.}
}
\value{
Lista con `w` y `b`.
}
\description{
Recupera pesos y sesgo desde un vector de estado theta.
}
\examples{
theta <- c(0.8, 0.3, 0.1)
rnas_unpack_neuron_params(theta, d_input = 2)

}
