% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{rnas_update_params_mlp}
\alias{rnas_update_params_mlp}
\title{Actualizar parametros de un MLP simple RNAS}
\usage{
rnas_update_params_mlp(params, grads, eta)
}
\arguments{
\item{params}{Lista con parametros actuales.}

\item{grads}{Lista con gradientes calculados por `rnas_mlp_backward()`.}

\item{eta}{Tasa de aprendizaje positiva.}
}
\value{
Lista de parametros actualizados.
}
\description{
Actualiza los parametros del MLP mediante descenso por gradiente.
}
\examples{
X <- matrix(c(0.5, -0.2,
              1.0,  0.3), ncol = 2, byrow = TRUE)
y <- c(0.4, 0.8)
params <- rnas_init_mlp(2, 3, seed = 123)
grads <- rnas_mlp_backward(X, y, params)
rnas_update_params_mlp(params, grads, eta = 0.1)

}
