% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrenamiento_neurona.R
\name{rnas_update_params_neuron}
\alias{rnas_update_params_neuron}
\title{Actualizar parámetros de una neurona RNAS}
\usage{
rnas_update_params_neuron(w, b, grad_w, grad_b, eta)
}
\arguments{
\item{w}{Vector numérico de pesos actuales.}

\item{b}{Sesgo numérico escalar actual.}

\item{grad_w}{Vector numérico de gradientes respecto a los pesos.}

\item{grad_b}{Gradiente numérico respecto al sesgo.}

\item{eta}{Tasa de aprendizaje positiva.}
}
\value{
Lista con `w_new` y `b_new`.
}
\description{
Actualiza pesos y sesgo usando la regla de descenso por gradiente:
w_new = w - eta * grad_w
b_new = b - eta * grad_b
}
\examples{
w <- c(0.8, 0.3)
b <- 0.1
grad_w <- c(-0.01, 0.02)
grad_b <- -0.03
rnas_update_params_neuron(w, b, grad_w, grad_b, eta = 0.1)

}
