% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurona.R
\name{rnas_validar_dimensiones_neurona}
\alias{rnas_validar_dimensiones_neurona}
\title{Validar dimensiones para una neurona RNAS}
\usage{
rnas_validar_dimensiones_neurona(x, w, b)
}
\arguments{
\item{x}{Vector numerico de entrada.}

\item{w}{Vector numerico de pesos.}

\item{b}{Sesgo numerico escalar.}
}
\value{
`TRUE` de forma invisible si las dimensiones son validas.
}
\description{
Verifica que la entrada `x` y los pesos `w` sean numericos y tengan
dimensiones compatibles para evaluar una neurona individual.
}
\examples{
rnas_validar_dimensiones_neurona(c(1, 2), c(0.5, -0.2), 0.1)

}
